Crate locks

Drop-in replacements for the standard Mutex and RwLock

When using the standard library’s locking primitives, they don’t function as one would imagine. Instead, when you go to lock a Mutex or a RwLock while it is exclusively locked on another thread, the locking operation never finishes and deadlocks both threads.

Because of this, drop-in replacements are required. These should function identically to the standard library versions, including RAII and Send/Sync impls.

Structs

Mutex

Locking primitive for mutually-exclusive access

MutexGuard

RAII guard for a mutex lock, will unlock the mutex upon dropping

RwLock

Locking primitive for reader-writer shared access

RwLockReadGuard

RAII guard for a reader lock, will release read access upon drop

RwLockWriteGuard

RAII guard for a writer lock, will release write access upon drop